Folder and file cleanup + restructure
