+++ apps/react/comments/website-builder/freestyle-comments/freestyle-comments-demo/components/document/FreestyleCanvas/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './FreestyleCanvas'
+export { default as FreestyleCanvas } from './FreestyleCanvas'
