+++ apps/react/comments/website-builder/freestyle-comments/freestyle-comments-demo/components/header/header.tsx
@@ -1,8 +1,11 @@
+"use client";
+
+import VeltTools from '../velt/VeltTools'; // [Velt] Component containing presence, notifications, huddle, and sidebar button tools
+
 export default function Header() {
   return (
-    <div className="flex items-center gap-2 p-2 border-b bg-background">
-      <h1 className="text-lg font-semibold">freestyle-comments-demo</h1>
-      {/* Add Velt components here: notifications, user presence, etc. */}
+    <div className="absolute top-2 right-6 flex items-center gap-[6px] z-50">
+      <VeltTools />
     </div>
-  )
+  );
 }
