+++ apps/react/comments/website-builder/freestyle-comments/freestyle-comments-demo/components/velt/VeltInitializeDocument.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { useEffect } from 'react';
-import { useSetDocuments } from '@veltdev/react';
+import { useSetDocuments, useCurrentUser } from '@veltdev/react';
 import { useCurrentDocument } from '@/app/document/useCurrentDocument';
 import { useAppUser } from '@/app/userAuth/useAppUser';
 
@@ -11,13 +11,16 @@ export default function VeltInitializeDocument() {
   // [Velt] Get document setter hook
   const { setDocuments } = useSetDocuments();
 
+  // [Velt] Wait for Velt user to be authenticated before setting document
+  const veltUser = useCurrentUser();
+
   // [Velt] Set document in Velt. This is the resource where all Velt collaboration data will be scoped.
   useEffect(() => {
-    if (!user || !documentId || !documentName) return;
+    if (!veltUser || !user || !documentId || !documentName) return;
     setDocuments([
       { id: documentId, metadata: { documentName: documentName || 'Untitled' } },
     ]);
-  }, [user, setDocuments, documentId, documentName]);
+  }, [veltUser, user, setDocuments, documentId, documentName]);
 
   return null;
 }
