+++ apps/react/comments/website-builder/freestyle-comments/freestyle-comments-demo/components/velt/VeltTools.tsx
@@ -6,33 +6,16 @@ import {
   VeltNotificationsTool,
 } from "@veltdev/react";
 
-/**
- * VeltTools - Top-right toolbar containing all Velt collaboration controls
- *
- * Component order (LEFT to RIGHT):
- * 1. Presence - Shows online users
- * 2. Freestyle Comments - VeltCommentTool for pin comments
- * 3. Text Comments - Enabled via textMode on VeltComments (auto-appears on text selection)
- * 4. Sidebar - Toggle comments sidebar
- * 5. Notifications - Notification panel
- * 6. Custom Annotation Dropdown - Configured in VeltCollaboration (appears in comment dialog)
- */
 function VeltTools() {
   return (
     <>
-      {/* 1. [Velt] Show online users - Presence */}
+      {/* [Velt] Show online users */}
       <VeltPresence />
-
-      {/* 2. [Velt] Freestyle Comments - Click to pin comments anywhere */}
+      {/* [Velt] Freestyle comments - click to pin comments anywhere */}
       <VeltCommentTool />
-
-      {/* 3. Text Comments - Enabled via textMode={true} on VeltComments */}
-      {/* Auto-appears when user selects text, no separate toolbar button needed */}
-
-      {/* 4. [Velt] Toggle comments sidebar */}
+      {/* [Velt] Toggle comments sidebar */}
       <VeltSidebarButton />
-
-      {/* 5. [Velt] Notifications panel */}
+      {/* [Velt] Notifications panel */}
       <VeltNotificationsTool
         settings={true}
         shadowDom={false}
@@ -42,9 +25,6 @@ function VeltTools() {
           all: { name: "All", enable: true },
         }}
       />
-
-      {/* 6. Custom Annotation Dropdown - Configured in VeltCollaboration */}
-      {/* Appears within the comment dialog, not as a toolbar button */}
     </>
   );
 }
